feature | create new username and password base component (#941)
* feature | create new username and password base component

* updates to enable easier extending

* fix backlinks on account details extended components
